Chat files code restructure
